moved to pipx

+++ tests/ansible_setup_local.sh
@@ -28,9 +28,6 @@ esac
 
 echo "INFO: determine OS version now"
 case "$VERSION_ID" in 
-  7*)
-    os_ver=7
-    ;;
   8*)
     os_ver=8
     ;;
@@ -40,22 +37,6 @@ case "$VERSION_ID" in
 esac
 
 echo "DEBUG: os_type=$os_type os_ver=$os_ver"
-
-if [ "$os_type$os_ver" == "rel_clone7"  ] #########################################################
-then
-  for PKG in epel-release git wget curl ansible
-  do
-     rpm -q $PKG >/dev/null 2>&1 && echo $PKG is already installed || yum -y install $PKG
-  done
-  ansibleconfigfile="/etc/ansible/ansible.cfg"
-  sed -i 's|^#inventory .*|inventory      = /etc/ansible/hosts|g' $ansibleconfigfile
-  sed -i 's|^#roles_path .*|roles_path    = /etc/ansible/roles|g' $ansibleconfigfile
-  sed -i 's|^#remote_user .*|remote_user = root|g' $ansibleconfigfile
-  sed -i 's|^#nocows .*|nocows = 1|g' $ansibleconfigfile
-  sed -i "/^roles_path/a\ \n#additional paths to search for collections in, colon separated\ncollections_paths = /etc/ansible/collections" $ansibleconfigfile
-  test -d /etc/ansible/projects || mkdir /etc/ansible/projects ; chmod 700 /etc/ansible/projects
-  test -d /etc/ansible/collections || mkdir /etc/ansible/collections ; chmod 755 /etc/ansible/collections
-  ansible-galaxy search joe-speedboat | cat
 
 elif [ "$os_type$os_ver" == "rel_clone8"  ] #########################################################
 then
@@ -124,24 +105,72 @@ then
 
 elif [ "$os_type$os_ver" == "rel_clone9"  ] #########################################################
 then
-  dnf -y config-manager --set-enabled crb
-  for PKG in https://dl.fedoraproject.org/pub/epel/epel-release-latest-9.noarch.rpm  https://dl.fedoraproject.org/pub/epel/epel-next-release-latest-9.noarch.rpm git wget curl ansible-core
-  do
-     rpm -q $PKG >/dev/null 2>&1 && echo $PKG is already installed || dnf -y install $PKG
-  done
-  ansibleconfigfile="/etc/ansible/ansible.cfg"
-  test -d /etc/ansible || mkdir /etc/ansible ; chmod 755 /etc/ansible
-  test -d /etc/ansible/projects || mkdir /etc/ansible/projects ; chmod 700 /etc/ansible/projects
-  test -d /etc/ansible/playbooks || mkdir /etc/ansible/playbooks ; chmod 700 /etc/ansible/playbooks
-  test -d /etc/ansible/collections || mkdir /etc/ansible/collections ; chmod 755 /etc/ansible/collections
-  test -f $ansibleconfigfile && cp -anv $ansibleconfigfile $ansibleconfigfile.bak
-  ansible-config init --disabled -t all | sed -e 's|{{ ANSIBLE_HOME ~ "/|/etc/ansible/|g' -e 's|" }}||g' > $ansibleconfigfile
-  sed -i 's|^;inventory=|inventory=|' $ansibleconfigfile
-  sed -i 's|^;roles_path=|roles_path=|' $ansibleconfigfile
-  sed -i 's|^;collections_path=|collections_path=|' $ansibleconfigfile
-  sed -i 's|^;nocows=.*|nocows=True|' $ansibleconfigfile
-  test -f /etc/ansible/hosts || echo localhost > /etc/ansible/hosts
-  ansible-galaxy search joe-speedboat | cat
+  # install system wide python
+  unset NPV ; PV=3.12 ; read -p  "Python version[$PV]: " NPV ; if [ "$NPV" ] ; then PV="$NPV" ; fi ; export PV="$PV"
+  unset AU ; AU=ansible ; read -p  "Ansible Username[$AU]: " NAU ; if [ "$NAU" ] ; then AU="$NAU" ; fi ; export AU="$AU"
+  unset SA ; SA=Y ; read -p  "Setup Ansible config in /etc/ansible[$SA]: " NSA ; if [ "$NSA" ] ; then SA="$NSA" ; fi ; export SA="$SA"
+
+  PKG="git wget curl python${PV} python${PV}-pip"
+  if [ "$USER" == "root" ]
+  then
+    echo INFO: User is root, installing $PKG
+    dnf -y install $PKG || exit 1
+
+    echo "INFO: Setup Ansible user: $AU"
+    id $AU >/dev/null 2>1 || useradd $AU
+    id $AU || exit 1
+
+    echo INFO: Installing Ansible as user $AU
+    su -s /bin/bash -c "pip$PV install pipx" $AU
+    su -s /bin/bash -c "~/.local/bin/pipx install --include-deps ansible" $AU
+
+    echo INFO: Test Ansible as user $AU
+    su -s /bin/bash -c "~/.local/bin/ansible --version" $AU
+
+    if [ "$SA" == "Y" ]
+    then
+      echo INFO: Setup Ansible devel env in /etc/ansible
+      ansibleconfigfile="/etc/ansible/ansible.cfg"
+      test -d /etc/ansible || mkdir /etc/ansible ; chmod 755 /etc/ansible
+      test -d /etc/ansible/projects || mkdir /etc/ansible/projects ; chmod 700 /etc/ansible/projects
+      test -d /etc/ansible/playbooks || mkdir /etc/ansible/playbooks ; chmod 700 /etc/ansible/playbooks
+      test -d /etc/ansible/collections || mkdir /etc/ansible/collections ; chmod 755 /etc/ansible/collections
+      test -d /etc/ansible/roles || mkdir /etc/ansible/roles ; chmod 755 /etc/ansible/roles
+      test -f $ansibleconfigfile && cp -anv $ansibleconfigfile $ansibleconfigfile.bak
+      chown -R $AU:$AU /etc/ansible
+
+      # Switch to user $AU
+      sudo -iu $AU <<EOF
+
+      . ~/.bashrc
+      # Initialize ansible configuration
+      ansibleconfigfile="/etc/ansible/ansible.cfg"
+      ansible-config init --disabled -t all | sed -e 's|{{ ANSIBLE_HOME ~ "/|/etc/ansible/|g' -e 's|" }}||g' > \$ansibleconfigfile
+      sed -i 's|^;inventory=|inventory=|' \$ansibleconfigfile
+      sed -i 's|^;roles_path=.*|roles_path=roles_path=/etc/ansible/roles|' \$ansibleconfigfile
+      sed -i 's|^;collections_path=.*|collections_path=/etc/ansible/collections|' \$ansibleconfigfile
+      sed -i 's|^;nocows=.*|nocows=True|' \$ansibleconfigfile
+
+      # Check if hosts file exists and configure localhost
+      test -f /etc/ansible/hosts || echo "localhost ansible_connection=local ansible_become=False" > /etc/ansible/hosts
+      # Test ansible shell module
+      ansible -m shell -a id localhost
+      echo "DONE, for testing role and collection overriding, run forex:
+
+      ansible-galaxy collection install fortinet.fortios:2.3.6
+      ansible-galaxy install joe-speedboat.os_update
+
+      as user $AU"
+EOF
+    else
+      echo SKIP: Setup Ansible devel env in /etc/ansible
+    fi
+    # Search galaxy for a collection
+    su -s /bin/bash -c "~/.local/bin/ansible-galaxy search joe-speedboat | cat" $AU
+  else
+    echo INFO: User is not root, we setup as root
+    exit 0
+  fi
 else
   echo "WARNING: No supported operating system found: os_type=$os_type os_ver=$os_ver"
 fi
